Tratando Exceções DELETE, PUT e POST

+++ src/main/java/com/dziombra/desafioCliente/services/ClientService.java
@@ -3,8 +3,11 @@ package com.dziombra.desafioCliente.services;
 import com.dziombra.desafioCliente.dto.ClientDTO;
 import com.dziombra.desafioCliente.entities.Client;
 import com.dziombra.desafioCliente.repositories.ClientRepository;
+import com.dziombra.desafioCliente.services.exceptions.DataBaseException;
 import com.dziombra.desafioCliente.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -39,15 +42,24 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update (Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete (Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        } else {
+            repository.deleteById(id);
+        }
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
